[1.2.4] fix many bugs

+++ src/main/java/com/hippo/fresh/entity/Product.java
@@ -58,11 +58,11 @@ public class Product {
     private Timestamp updateTime;//商品更新时间
 
     @Column
-    private int commentNum;//评论数
+    private Long commentNum;//评论数
 
     @Column
-    private int  commentScore;//总评分
+    private Long  commentScore;//总评分
 
     @Column
-    private double score;//评分
+    private Double score;//评分
 }
